Fix player creation server module exports and API auth checks
