+++ src/app/api/players/route.ts
@@ -3,9 +3,11 @@ import type { Role } from "@prisma/client";
 
 import { auth } from "@/lib/auth";
 import { prisma } from "@/lib/prisma";
-
 import { hasPermission, PERMISSIONS } from "@/lib/rbac";
-import { createPlayerSchema, normalizePlayerInput } from "@/lib/players";
+import {
+  createPlayerSchema,
+  normalizePlayerInput,
+} from "@/lib/players";
 import type { ZodIssue } from "@/lib/zod";
 
 type Session = Awaited<ReturnType<typeof auth>>;
@@ -28,30 +30,26 @@ function buildFieldErrors(issues: ZodIssue[]): FieldErrors {
 
 function getAuthorizedUser(
   session: Session,
-  permission: keyof typeof PERMISSIONS
+  permission: (typeof PERMISSIONS)[keyof typeof PERMISSIONS]
 ) {
   const role = session?.user?.role as Role | undefined;
   const userId = session?.user?.id;
+
   if (!role || !userId) {
     return null;
   }
 
-  if (!hasPermission(role, permission)) {
+  if (!hasPermission(role, permission as keyof typeof PERMISSIONS)) {
     return null;
   }
 
   return { userId, role };
 }
 
-import { persistPlayer, createPlayerSchema } from "@/app/players/actions";
-import { hasPermission, PERMISSIONS, ROLES } from "@/lib/rbac";
-
-const ROLE_VALUES = Object.values(ROLES) as Role[];
-
-
 export async function GET() {
   const session = await auth();
   const authUser = getAuthorizedUser(session, PERMISSIONS["players:read"]);
+
   if (!authUser) {
     return NextResponse.json({ error: "Unauthorized" }, { status: 401 });
   }
@@ -73,8 +71,9 @@ export async function GET() {
 }
 
 export async function POST(req: NextRequest) {
-
+  const session = await auth();
   const authUser = getAuthorizedUser(session, PERMISSIONS["players:create"]);
+
   if (!authUser) {
     return NextResponse.json({ error: "Unauthorized" }, { status: 401 });
   }
@@ -94,69 +93,17 @@ export async function POST(req: NextRequest) {
     const fieldErrors = buildFieldErrors(parsed.error.issues);
     return NextResponse.json(
       { error: "Validation échouée.", fieldErrors },
-
-
-  if (!session?.user?.id) {
-    return NextResponse.json({ error: "Unauthorized" }, { status: 401 });
-  }
-
-  const role =
-    session.user.role && ROLE_VALUES.includes(session.user.role as Role)
-      ? (session.user.role as Role)
-      : null;
-
-  if (!role || !hasPermission(role, PERMISSIONS["players:create"])) {
-    return NextResponse.json({ error: "Forbidden" }, { status: 403 });
-  }
-
-  let payload: unknown;
-  try {
-    payload = await req.json();
-  } catch {
-    return NextResponse.json({ error: "Invalid JSON" }, { status: 400 });
-  }
-
-  const body =
-    typeof payload === "object" && payload !== null
-      ? (payload as Record<string, unknown>)
-      : {};
-
-  const candidate = {
-    name:
-      typeof body.name === "string"
-        ? body.name.trim()
-        : body.name,
-    position:
-      typeof body.position === "string"
-        ? body.position.trim()
-        : body.position,
-  };
-
-  const parsed = createPlayerSchema.safeParse(candidate);
-
-  if (!parsed.success) {
-    const fieldErrors: Record<string, string[]> = {};
-    for (const issue of parsed.error.issues) {
-      const field = issue.path[0];
-      if (typeof field === "string") {
-        fieldErrors[field] = [...(fieldErrors[field] ?? []), issue.message];
-      }
-    }
-
-    return NextResponse.json(
-      { error: "Invalid payload", fieldErrors },
-
       { status: 400 }
     );
   }
 
-  const data = normalizePlayerInput(parsed.data);
+  const normalized = normalizePlayerInput(parsed.data);
 
   try {
     const player = await prisma.player.create({
       data: {
-        name: data.name,
-        position: data.position,
+        name: normalized.name,
+        position: normalized.position.toUpperCase(),
         creatorId: authUser.userId,
       },
       select: { id: true, name: true, position: true, createdAt: true },
@@ -167,15 +114,6 @@ export async function POST(req: NextRequest) {
     console.error("[Players] Failed to create player", error);
     return NextResponse.json(
       { error: "Impossible d'enregistrer ce joueur." },
-
-  try {
-    const player = await persistPlayer(parsed.data, session.user.id);
-    return NextResponse.json(player, { status: 201 });
-  } catch (error) {
-    console.error("Failed to create player", error);
-    return NextResponse.json(
-      { error: "Unable to create player" },
-
       { status: 500 }
     );
   }
